backend/requirement.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
langchain
langchain-groq
langchain-community
langchain-core
langchain-google-genai
langgraph
pandas
pypdf
iso4217
PyMuPDF
faiss-cpu
langchain_mistralai
google-generativeai
python-multipart
matplotlib
seaborn
scikit-learn
tiktoken
sentence-transformers
langchain-huggingface